requirements.txt
sentence-transformers
faiss-cpu
numpy
dashscope
edge-tts
streamlit
pyyaml
requests
beautifulsoup4
pillow
